Make `VMAJavaTraceUtil` optional, enabled in opt-in manner
